Refactoring: visitor pattern in Selector
Commands can run on windows or desktops, but typically not both. Add a
`runWindow` method to `Selector`, which takes a function and runs that
function over each window (`runWindow :: (Window -> IO ()) -> IO ()`).
It's up to the selector to get the list of windows to run over. The
command classes, in turn, use `runWindow` instead of doing their own
iteration.

There is not currently a `runDesktop` method because I do not yet
implement any virtual desktop commands.

This also introduces a `WindowCommand` class that other command classes
are welcome to inherit from. It takes care of most of the parts that
change the least.