Fail Quitely on Parsing Error
Orginal plan was to allow comments to start at the beginning of a line,
but didn't figure out how to let the parser respect that. Instead,
runner still works if there are no command and selectors parsed. This
fixes the beginning of line comments and other previously bad input such
that nothing happens. It seems nice that it doesn't crash as much, but
maybe failing silently is not the way to go and more verbouse error
handling would be would make more sense.