Install the wim(7) manpage
This uses a slight hack, inpsired by https://github.com/novel/lc-tools/ .

Add a custom installer subclass that runs a script to install the man
pages.

Man pages are installed to `$PREFIX/man` or, if that doesn't exist, to
`$PREFIX/share/man`.

Remove the wim-grammer text file because that is now part of the
installed manpage.