Introduce a GTK+ GUI
The command-line UI was giving us threading issues. Resolve that by
switching to a GTK+ GUI instead.

The GUI sits at the bottom of the screen, full width. It has a simple
GtkEntry in it, and when you press enter it runs the command.

Rename wimsh to wim-gtk, in celebration.

This simplifies some existing problems: remove the InteractiveWim class
and remove the `Model.shutdown()` method.

We capture ^C, which is tricky, using the method described here:
https://bugzilla.gnome.org/show_bug.cgi?id=622084

"Fixes threading issues!" - Rebecca