Show user Errors instead of printing to shell.
All errors that are meant to be shown to the user are throw as
as the custom Exception class WimException. All these errors are caught and
printed to the user along with their orginal command an error icon. The user can
then clear the entry themselves and the next time they run a successful command
the error icon will be removed.

The only issue is that it is not super user friendly to have to clear the whole
entry to write your new command. Maybe the error should be cleared the next time
the user moves there cursor to wim.

Added a style guide to DEVELOPERS.md. Use PEP8 added a defined structure for the
imports. I fixed the existing files imports to follow this structure.