Move all Wnck calls to WnckWrapper
All calls to the Wnck objects are called via call_window, call_workspace,
call_application, call_screen, or call_class_group by passing in the
name of the method and the argument.

The WnckState contained by the WnckWrapper is now private as are all the
properties of the state that used to be directly exposed.